:recycle: rename to setting_page
