+++ lib/domain/home/home_page.dart
@@ -4,7 +4,7 @@ import 'package:pilll/database/database.dart';
 import 'package:pilll/domain/calendar/calendar_page.dart';
 import 'package:pilll/domain/menstruation/menstruation_page.dart';
 import 'package:pilll/domain/record/record_page.dart';
-import 'package:pilll/domain/settings/settings_page.dart';
+import 'package:pilll/domain/settings/setting_page.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/text_color.dart';
 import 'package:pilll/service/push_notification.dart';
